daggerize: reorganize
